Item fetching refactor
