+++ tests/dummy/app/controllers/media-registry/cardflow.js
@@ -6,6 +6,8 @@ import { fetchCollection } from 'dummy/media';
 
 export default class MediaRegistryCardflowController extends Controller {
   @tracked isolatedCollection = this.getIsolatedCollection(this.catalog.id);
+  @tracked itemId = null;
+  @tracked record = null;
 
   catalog = {
     id: 'batch-f',
@@ -21,6 +23,17 @@ export default class MediaRegistryCardflowController extends Controller {
       "media-registry/covers/thumb/Love-Never-Dies.jpg",
       "media-registry/covers/thumb/Animals.jpg"
     ]
+  }
+
+  @action
+  setItemId(item) {
+    if (item) {
+      this.itemId = dasherize(item.song_title.trim());
+      this.getRecord();
+    } else {
+      this.itemId = null;
+      this.record = null;
+    }
   }
 
   @action
@@ -87,5 +100,45 @@ export default class MediaRegistryCardflowController extends Controller {
         },
       ],
     };
+  }
+
+  @action
+  async getRecord() {
+    let itemId = this.itemId;
+    if (!itemId) { return; }
+
+    const records = await fetchCollection('all_tracks_combined');
+    const recordDetails = await fetchCollection('songs_by_pia_midina_bb_clarke_table_1');
+    const profiles = await fetchCollection('profiles');
+    const musicalWorks = await fetchCollection('musical-works');
+
+
+    const record = records.filter(item => {
+      if (item.catalog) {
+        return dasherize(item.song_title.trim()) === itemId;
+      }
+    })[0];
+
+    const recordDetail = recordDetails.filter(item => {
+      return dasherize(item.song_title.trim()) === itemId;
+    })[0];
+
+    const artist = profiles.filter(profile => {
+      return profile.id === dasherize(record.artist.trim());
+    })[0];
+
+    if (artist) {
+      record.artist_info = artist;
+    }
+
+    if (recordDetail) {
+      const producer = profiles.filter(profile => (profile.id === recordDetail.producer_id))[0];
+      const musicalWork = musicalWorks.filter(item => item.iswc === recordDetail.iswc_id)[0];
+      record.producer = producer;
+      record.details = recordDetail;
+      record.musicalWork = musicalWork;
+    }
+
+    this.record = record;
   }
 }
