+++ tests/dummy/app/templates/media-registry/cardflow.hbs
@@ -2,4 +2,7 @@
   @model={{@model}}
   @isolatedCatalog={{this.isolatedCollection}}
   @catalog={{this.catalog}}
+  @setItemId={{this.setItemId}}
+  @itemId={{this.itemId}}
+  @item={{this.record}}
 />
